+++ inputmask/src/main/kotlin/com/redmadrobot/inputmask/model/CaretString.kt
@@ -7,4 +7,20 @@ package com.redmadrobot.inputmask.model
  *
  * @author taflanidi
  */
-data class CaretString(val string: String, val caretPosition: Int)
+data class CaretString(
+    val string: String,
+    val caretPosition: Int,
+    val caretGravity: CaretGravity = CaretGravity.FORWARD
+) {
+    fun reversed() =
+        CaretString(
+            this.string.reversed(),
+            this.string.length - this.caretPosition,
+            this.caretGravity
+        )
+
+    enum class CaretGravity {
+        FORWARD,
+        BACKWARD
+    }
+}